passing variables

+++ transfer/docbook-production.sh
@@ -11,10 +11,17 @@ echo CHROOT=${CHROOT}
 hostname -f
 if [ "X${CHROOT}" != "X" ] ; then
 echo CHROOT=${CHROOT}
-chroot ${CHROOT}
+ycheckrc chroot ${CHROOT}
 fi
 
-
+function ycheckrc () {
+echo $@
+$@
+rc=$?
+if [ $[${rc}] != 0 ] ; then
+    exit ${rc}
+fi
+}
 SVNLOCATION="https://svnsrv.desy.de/public/grid-virt/org.hepix.virtulization.docs/trunk"
 #TAG=`svn ls ${SVNLOCATION} | org_desy_grid_virt_sort_release.py | tail -n 1`
 rm -rf build
